app/requirements.txt
tls-client~=1.0.1
fastapi~=0.104.1
html5lib~=1.1
python-dotenv~=1.0.0
tomd~=0.1.3
pandoc~=2.3
python-decouple~=3.8
pypandoc~=1.12
prompt_toolkit~=3.0.41
rich~=13.6.0
uvicorn~=0.24.0.post1
pika~=1.3.2
markdownify~=0.11.6
newspaper3k~=0.2.8
pydantic~=2.5.1
pysocks~=1.7.1
requests~=2.27.0
spacy~=3.7.2
spacy-html-tokenizer~=0.1.3
nltk~=3.8.1
textblob~=0.17.1
cachetools~=5.3.2
sqlalchemy~=1.4.25
aiosqlite~=0.20.0
